Refactor video out
