+++ libvideolive.cpp
@@ -7,7 +7,6 @@
 // See README for license
 
 #include <Python.h>
-#include <fcntl.h>
 #include <string.h>
 #include <string>
 #include <sstream>
@@ -15,7 +14,6 @@
 #include <vector>
 #include <stdexcept>
 #include <pthread.h>
-#include <dirent.h>
 #include "pixfmt.h"
 #include "v4l2capture.h"
 #include "videoout.h"
@@ -292,24 +290,13 @@ static PyObject *Device_manager_close(Device_manager *self, PyObject *args)
 }
 
 static PyObject *Device_manager_list_devices(Device_manager *self)
-{
+{	
 	PyObject *out = PyList_New(0);
-	const char dir[] = "/dev";
-	DIR *dp;
-	struct dirent *dirp;
-	if((dp  = opendir(dir)) == NULL) {
-		printf("Error(%d) opening %s\n", errno, dir);
-		Py_RETURN_NONE;
+	std::vector<std::string> devLi = List_in_devices();
+	for(unsigned i=0; i<devLi.size(); i++)
+	{
+		PyList_Append(out, PyString_FromString(devLi[i].c_str()));
 	}
-
-	while ((dirp = readdir(dp)) != NULL) {
-		if (strncmp(dirp->d_name, "video", 5) != 0) continue;
-		std::string tmp = "/dev/";
-		tmp.append(dirp->d_name);
-		PyList_Append(out, PyString_FromString(tmp.c_str()));
-	}
-	closedir(dp);
-
 	PyList_Sort(out);
 	return out;
 }
