+++ v4l2out.cpp
@@ -13,6 +13,8 @@
 #include <time.h>
 #include <dirent.h>
 #include <vector>
+#include <errno.h>
+#include <dirent.h>
 #include <stdexcept>
 #include "v4l2out.h"
 #include "pixfmt.h"
@@ -340,4 +342,27 @@ void *Video_out_manager_Worker_thread(void *arg)
 }
 
 // *****************************************************************
+
+std::vector<std::string> List_out_devices()
+{
+	std::vector<std::string> out;
+	const char dir[] = "/dev";
+	DIR *dp;
+	struct dirent *dirp;
+	if((dp  = opendir(dir)) == NULL) {
+		printf("Error(%d) opening %s\n", errno, dir);
+		return out;
+	}
+
+	while ((dirp = readdir(dp)) != NULL) {
+		if (strncmp(dirp->d_name, "video", 5) != 0) continue;
+		std::string tmp = "/dev/";
+		tmp.append(dirp->d_name);
+		out.push_back(tmp);
+	}
+	closedir(dp);
+	return out;
+}
+
+
